🤖 refactor(site/src/pages): nest workspace settings pages
