+++ site/src/pages/WorkspacePage/WorkspaceSettingsPage/WorkspaceSharingPage/WorkspaceSharingPageView.stories.tsx
@@ -62,7 +62,8 @@ const aclWithUsersAndGroups: WorkspaceACL = {
 };
 
 const meta: Meta<typeof WorkspaceSharingPageView> = {
-	title: "pages/WorkspaceSharingPageView",
+	title:
+		"pages/WorkspacePage/WorkspaceSettingsPage/WorkspaceSharingPage/WorkspaceSharingPageView",
 	component: WorkspaceSharingPageView,
 	parameters: {
 		queries: [
