+++ site/src/pages/WorkspacePage/WorkspaceSettingsPage/WorkspaceSharingPage/WorkspaceSharingPageView.tsx
